Jobs: Added FloatField to sort the results on leaderboard

+++ apps/jobs/views.py
@@ -5,7 +5,7 @@ from rest_framework.decorators import (api_view,
                                        throttle_classes,)
 
 from django.db.models.expressions import RawSQL
-from django.db.models import IntegerField
+from django.db.models import FloatField
 
 from rest_framework_expiring_authtoken.authentication import (
     ExpiringTokenAuthentication,)
@@ -140,7 +140,7 @@ def leaderboard(request, challenge_phase_split_id):
     leaderboard_data = LeaderboardData.objects.filter(
         challenge_phase_split=challenge_phase_split).order_by('created_at')
     leaderboard_data = leaderboard_data.annotate(
-        filtering_score=RawSQL('result->>%s', (default_order_by, ), output_field=IntegerField())).values(
+        filtering_score=RawSQL('result->>%s', (default_order_by, ), output_field=FloatField())).values(
             'id', 'submission__participant_team__team_name',
             'challenge_phase_split', 'result', 'filtering_score', 'leaderboard__schema')
